feat: move template creation to header
