+++ kanban-app/src/app/pages/main/main.component.html
@@ -16,10 +16,6 @@
         <mat-card-title>{{ 'BoardsStats' | translate }}</mat-card-title>
       </mat-card-header>
       <mat-card-content class="statistics-list">
-        <h4 class="basket" (click)="createCustomBoard()">
-          <mat-icon class="statistics-icon">create_new_folder</mat-icon>
-          <span class="statistics-text">{{ 'Create custom' | translate }}</span>
-        </h4>
         <h4 class="boards-all">
           <mat-icon class="statistics-icon">dashboard</mat-icon
           ><span class="statistics-text"
