+++ kanban-app/src/assets/i18n/by.json
@@ -144,5 +144,7 @@
   "View": "Выгляд: ",
   "Max": "пашыраны",
   "Min": "скарочаны",
-  "ChoosePicture": "Абярыце малюнак"
+  "ChoosePicture": "Абярыце малюнак",
+  "blankBoard": "Пустая дошка",
+  "templateBoard": "Дошка па шаблоне"
 }
