+++ kanban-app/src/assets/i18n/ru.json
@@ -143,5 +143,7 @@
   "View": "Вид: ",
   "Max": "расширенный",
   "Min": "сокращенный",
-  "ChoosePicture": "Выберите изображение"
+  "ChoosePicture": "Выберите изображение",
+  "blankBoard": "Пустая доска",
+  "templateBoard": "Доска по шаблону"
 }
