+++ kanban-app/src/assets/i18n/ua.json
@@ -143,5 +143,7 @@
   "View": "Вид: ",
   "Max": "розширений",
   "Min": "скорочений",
-  "ChoosePicture": "Оберіть зображення"
+  "ChoosePicture": "Оберіть зображення",
+  "blankBoard": "Порожня дошка",
+  "templateBoard": "Дошка за шаблоном"
 }
